Move opml_urlreader into separate set of files
