+++ mk/mk.deps
@@ -82,12 +82,13 @@ src/controller.o: src/controller.cpp include/controller.h include/cache.h \
  include/feedhq_api.h include/file_urlreader.h include/formatstring.h \
  include/globals.h include/inoreader_api.h include/logger.h \
  include/newsblur_api.h rss/rsspp.h include/remote_api.h \
- include/ocnews_api.h include/oldreader_api.h include/regexmanager.h \
- include/rss_parser.h include/stflpp.h include/strprintf.h \
- include/ttrss_api.h 3rd-party/json.hpp include/utils.h include/view.h \
- include/controller.h include/filebrowser_formaction.h \
- include/formaction.h include/history.h include/keymap.h include/stflpp.h \
- include/htmlrenderer.h include/textformatter.h
+ include/ocnews_api.h include/oldreader_api.h include/opml_urlreader.h \
+ include/regexmanager.h include/rss_parser.h include/stflpp.h \
+ include/strprintf.h include/ttrss_api.h 3rd-party/json.hpp \
+ include/utils.h include/view.h include/controller.h \
+ include/filebrowser_formaction.h include/formaction.h include/history.h \
+ include/keymap.h include/stflpp.h include/htmlrenderer.h \
+ include/textformatter.h
 src/dialogs_formaction.o: src/dialogs_formaction.cpp \
  include/dialogs_formaction.h include/formaction.h include/history.h \
  include/keymap.h include/configparser.h include/rss.h \
@@ -153,7 +154,8 @@ src/filebrowser_formaction.o: src/filebrowser_formaction.cpp \
  include/opml.h include/filebrowser_formaction.h include/htmlrenderer.h \
  include/textformatter.h
 src/file_urlreader.o: src/file_urlreader.cpp include/file_urlreader.h \
- include/urlreader.h include/configcontainer.h include/configparser.h
+ include/urlreader.h include/utils.h include/configcontainer.h \
+ include/configparser.h include/logger.h config.h include/strprintf.h
 src/filtercontainer.o: src/filtercontainer.cpp include/filtercontainer.h \
  include/configparser.h config.h include/exceptions.h include/matcher.h \
  filter/FilterParser.h include/strprintf.h include/utils.h \
@@ -293,6 +295,8 @@ src/opml.o: src/opml.cpp include/opml.h include/feedcontainer.h \
  include/rss.h include/configcontainer.h include/configparser.h \
  include/matcher.h filter/FilterParser.h include/utils.h include/logger.h \
  config.h include/strprintf.h include/urlreader.h
+src/opml_urlreader.o: src/opml_urlreader.cpp include/opml_urlreader.h \
+ include/configcontainer.h include/configparser.h include/urlreader.h
 src/pb_controller.o: src/pb_controller.cpp include/pb_controller.h \
  include/configcontainer.h include/configparser.h include/download.h \
  include/fslock.h include/queueloader.h include/pb_controller.h \
@@ -405,9 +409,7 @@ src/ttrss_urlreader.o: src/ttrss_urlreader.cpp include/ttrss_api.h \
  include/strprintf.h include/remote_api.h rss/rsspp.h \
  include/remote_api.h include/urlreader.h include/file_urlreader.h \
  include/logger.h
-src/urlreader.o: src/urlreader.cpp include/urlreader.h \
- include/configcontainer.h include/configparser.h config.h \
- include/logger.h include/strprintf.h include/utils.h include/logger.h
+src/urlreader.o: src/urlreader.cpp include/urlreader.h
 src/urlview_formaction.o: src/urlview_formaction.cpp \
  include/urlview_formaction.h include/formaction.h include/history.h \
  include/keymap.h include/configparser.h include/rss.h \
@@ -466,8 +468,7 @@ test/feedcontainer.o: test/feedcontainer.cpp 3rd-party/catch.hpp \
  include/logger.h config.h include/strprintf.h include/configcontainer.h \
  include/feedcontainer.h include/rss.h test/test-helpers.h
 test/file_urlreader.o: test/file_urlreader.cpp include/file_urlreader.h \
- include/urlreader.h include/configcontainer.h include/configparser.h \
- 3rd-party/catch.hpp test/test-helpers.h
+ include/urlreader.h 3rd-party/catch.hpp test/test-helpers.h
 test/filterparser.o: test/filterparser.cpp filter/FilterParser.h \
  3rd-party/catch.hpp
 test/formatstring.o: test/formatstring.cpp include/formatstring.h \
@@ -527,8 +528,7 @@ test/textformatter.o: test/textformatter.cpp include/textformatter.h \
  include/regexmanager.h include/configparser.h include/matcher.h \
  filter/FilterParser.h 3rd-party/catch.hpp
 test/urlreader.o: test/urlreader.cpp include/urlreader.h \
- include/configcontainer.h include/configparser.h 3rd-party/catch.hpp \
- test/test-helpers.h
+ 3rd-party/catch.hpp test/test-helpers.h
 test/utils.o: test/utils.cpp include/utils.h include/configcontainer.h \
  include/configparser.h include/logger.h config.h include/strprintf.h \
  3rd-party/catch.hpp
